feat(023): engage delay, history reset, float variationScale, test3/4 analysis
Core changes:
- AircraftState::resetHistory(targetPos, pathTangent): universal history
  pre-fill method used by both CRRCSim and minisim at scenario start.
  Pre-fills all 10 slots with real geometry so NN sees consistent
  direction cosines from tick 1 (not zeros).

- computeVariationScale() returns float (was double): fixes eval
  determinism bug where double→float truncation of 6/9=0.6667 caused
  ~2e-8 delta in rabbitSpeedConfig.sigma and variation offsets. Rule:
  all values that round-trip through serialization must be float.
  Verified: test4 eval returns NN_EVAL_SAME.

- CRRCSim submodule bump: engage delay window (750ms), history pre-fill,
  initial throttle from entrySpeedFactor.

- autoc.ini: pop 3500 (2x params), posR/posAlt=0 (origin start).
  test4 at pop 3500 tracks test3 (pop 10k) convergence at 3x speed.

Analysis & planning:
- sim_polar_viz.py: updated for 023 direction cosines, stepPoints/mult
  panel, prefiltered input workflow.
- Fitness curves: test2 vs test3 vs test4 comparison plots.
- nn_dirvec path visualizations from test3 gen 260.
- remaining-work.md: prioritized 023 wrap-up plan (pt3 filter, rabbit
  speed, memory investigation, US1/US2 test coverage).
- 024 spec scaffold: craft variations, trim offsets, authority limit.
- Clarifications session 2026-04-12: pt3 filter design decisions.

+++ src/autoc.cc
@@ -75,17 +75,18 @@ static float gEvalVariationScaleOverride = -1.0f;  // >=0 = use this instead of 
  *
  * @return Scale factor to apply to variation offsets (0.0 to 1.0)
  */
-static double computeVariationScale() {
+static float computeVariationScale() {
     // Eval mode: use the exact scale stored in the weight file
     if (gEvalVariationScaleOverride >= 0.0f) return gEvalVariationScaleOverride;
 
-    // Ramp from 0.0 (no variations) to 1.0 (full variations) over training.
-    // Disabled (rampStep<=0), eval mode (totalGens<=1), or trivial: return 1.0.
+    // Ramp from 0.0 to 1.0 over training.  Returns float for consistency with
+    // serialized variation_scale (stored as float in NN01 weights).  All call
+    // sites must see identical precision between training and eval.
     int numSteps = (gVariationRampStep > 0) ? gTotalGenerations / gVariationRampStep : 0;
-    if (numSteps <= 1) return 1.0;
+    if (numSteps <= 1) return 1.0f;
 
     int stepIndex = (gCurrentGeneration - 1) / gVariationRampStep;  // 1-based gen
-    return static_cast<double>(std::min(stepIndex, numSteps - 1)) / static_cast<double>(numSteps - 1);
+    return static_cast<float>(std::min(stepIndex, numSteps - 1)) / static_cast<float>(numSteps - 1);
 }
 
 // ============================================================================
@@ -249,7 +250,7 @@ static void populateVariationOffsets(ScenarioMetadata& meta) {
     const auto& v = gScenarioVariations[windIdx].entryOffsets;
 
     // Get current scale (0.0 to 1.0 over training)
-    double scale = computeVariationScale();
+    float scale = computeVariationScale();
 
     // Angular offsets: scale toward 0
     meta.entryHeadingOffset = v.entryHeadingOffset * scale;
